extract pmutil, and squash commits from gitlab
This is a huge change.

 - stream is supported

For `impl Stream<Item = T, Error = E>`,
 yield type is `Result<T, StreamError<E>>` and return type is `()`.
 This signature reflects the fact that stream from futures v0.1
 does not support terminating itself with error, and if it's
 supported in future, fn signature will be changed.
StreamError is a internal error type which has `NotReady` variant.

 - non-static lifetime is supported

This is implemented by prepending a special
 lifetime `'__returned_future` and bounding returned future like
 `impl Future + '__returned_future`.
But this should **not** be done if signature matters
 (e.g. inside trait impl).
Opt out is not implemented yet.

 - yield / return types are explicitly annotated

So rustc can use them for type inference and error reporting.

 - await! macro works for different type of futures

As user get `Result<T, E>` from `await!` regardless of future provided,
 `await!(if a { fut_a } else { fut_b })` expands to
 `if a { await!(fut_a) } else { await!(fut_b) }` and
 give `Result<T, E>` back to user.